separate QuadraticProgram class to another file because it should be considered as a separate step
